Add setting to optionally split at the end of the intro section
The easiest way to detect the "button smash" at the end of the intro is by comparing the current player position with (-10152.54, -2685.575, 3148.311). As the player position struct location in memory changes when the game is loading a map, the respective pointer path needs to be dereferenced every time this happens.

At the moment the optional splits are only tested and thus supported for the latest version of the game on Steam.